requirements.txt
playwright
requests
python-dotenv
tzdata
python-telegram-bot[job-queue]